Moved DatabaseConnector
